+++ galaxy-dal/src/main/java/com/wuxiu/galaxy/dal/domain/LuggageStorageRecord.java
@@ -10,24 +10,25 @@ package com.wuxiu.galaxy.dal.domain;
 
 import com.baomidou.mybatisplus.annotations.TableField;
 import com.baomidou.mybatisplus.annotations.TableId;
+import com.baomidou.mybatisplus.annotations.TableLogic;
 import com.baomidou.mybatisplus.annotations.TableName;
 import com.baomidou.mybatisplus.enums.IdType;
 import com.wuxiu.galaxy.api.common.base.BaseModel;
-import com.baomidou.mybatisplus.annotations.*;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 import java.time.LocalDateTime;
 
-/**  
- * <p>LuggageStorageRecordDomian实体对象</p>
- *
- * 行李寄存-取件表
- *
- * @author: Baomidou_Generater（rarestzhou@gmail.com）
- * @since 2019-04-15
- */
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
+/**
+ *   
+ *  * <p>LuggageStorageRecordDomian实体对象</p>
+ * <p>
+ *  行李寄存-取件表
+ * <p>
+ *  * @author: Baomidou_Generater（rarestzhou@gmail.com）
+ *  * @since 2019-04-15
+ *  
+ */
 
 @Data
 @EqualsAndHashCode(callSuper = true)
@@ -44,7 +45,7 @@ public class LuggageStorageRecord extends BaseModel {
      * 行李类型主键id
      */
     @TableField("luggage_type_id")
-    private Integer luggageTypeId;
+    private Long luggageTypeId;
     /**
      * 管理员id
      */
